IndexerIntegrationTest: Fix method order
IndexerIntegerationTest is annotated with JUnit 4's
org.junit.FixMethodOrder, and is intended to have its tests executed
in an alphabetical order.

However, the JUnit Jupiter test runner does not honor this annotation,
and the order of execution is not alphabetical. Given that the test
passes regardless of the order and this annotation is not honored
anyway, it can just be removed.