Rewrite KubernetesTest as a JUnit Jupiter exception
By instantiating an
io.fabric8.kubernetes.client.server.mock.KubernetesServer object, the
KubernetesTest implicitly using JUnit 4 classes under the hood.

This patch replaces that class with the modern
@EnableKubernetesMockClient extension, allowing for a fully JUnit
Jupiter runtime. As a side bonus, so of the boilerplate code
associated with a KubernetesServer is removed, and the test is cleaner
and easier to maintain.